requirements.txt
flask==2.3.3
openai==1.14.2
python-dotenv==1.0.1